+++ src/pyramid/view.py
@@ -20,9 +20,14 @@ from pyramid.interfaces import (
     IViewClassifier,
 )
 from pyramid.threadlocal import get_current_registry, manager
-from pyramid.util import hide_attrs, reraise as reraise_
+from pyramid.util import Sentinel, hide_attrs, reraise as reraise_
 
 _marker = object()
+
+if sys.version_info.major < 3 or sys.version_info.minor < 11:
+    Self = Sentinel('Self')  # pragma: no cover
+else:
+    from typing import Self  # noqa: F401
 
 
 def render_view_to_response(context, request, name='', secure=True):